\name{display-methods}
\docType{methods}
\alias{display-methods}
\alias{display,orderbook-method}
\alias{display}

% Need this last even thought promptMethod() does not provide it by
% default.

\title{Method for Function display in Package `orderbook'}
\description{
  Methods for function \code{display} in Package `orderbook' ~~
}
\section{Methods}{
\describe{
  \item{\code{signature(object = "orderbook")}}{
    Prints by default the best 5 price levels and corresponding sizes for the bid
    and ask in the orderbook.
  }
}
}
\keyword{methods}

