\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,orderbook-method}
\alias{show}
\title{Methods for Function show for Package `orderbook'}
\description{
Methods for function \code{show} in Package `orderbook'
}
\section{Methods}{
\describe{

  \item{\code{signature(object = "orderbook")}}{
    Simple print method for orderbook objects.
  }
}}
\keyword{methods}
