\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary,orderbook-method}
\alias{summary}
\title{Method for Function summary in Package `orderbook'}
\description{
  Method for function \code{summary} in Package `orderbook'.
}
\section{Methods}{
\describe{
  \item{\code{signature(object = "orderbook")}}{
    Prints a summary of the \code{orderbook} object.
  }
}}
\keyword{methods}

