\name{orderbook-package}
\alias{orderbook-package}
\docType{package}
\title{
Analyze and visualize orderbook data.
}
\description{
The orderbook package provides functions for exploring and visualizing
orderbook data. For example, we may want to study an orderbook
immediately before a trade occurs, or maybe 50 orders before a trade
occurs, or we want to visualize the change of an orderbook overtime as
the price level fluctuates. The insights from this sort
of analysis can help in researching short-term trading strategies.
}
\details{
\tabular{ll}{
Package: \tab orderbook\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-07-01\cr
Depends: \tab R (>= 2.11.0), hash, methods, grid, lattice\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

Index:
\preformatted{
orderbook			Creating an Object of class orderbook
orderbook-class			Class "orderbook"
}

Further information is available in the following vignettes:
\tabular{ll}{
\code{orderbook} \tab Using the orderbook package (source, pdf)\cr
}
}
\author{
Andrew Liu <andrew.t.liu@williams.edu>, Khanh Nguyen
<knguyen@cs.umb.edu>, David Kane <dave@kanecap.com>.
}
\keyword{package}
